Huge refacto
For once, the code is structured into functions, and doesn't have any
globals anymore.
I also introduced a short Jupyter notebook for plots.

The procedural code is still here in main.py, to be used by Latex.

For now, these functions are not in the right modules, I will move them
later on.